+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -22,3 +22,4 @@ CHANGELOG
  * [BC BREAK] collections in fields annotated with `Valid` are not traversed
    recursively anymore by default. `Valid` contains a new property `deep`
    which enables the BC behavior.
+ * added MinCount and MaxCount constraint